Refactor main tests to remove duplication
Originally this test served as an integration test that helped drive the creation of `todoist.GetTask()`. Now that’s done this was overlapping in responsibility.

This test now asserts that when the function is called with a client and string, the client calls getTasks.

We will need to move the old logging behaviour to get tasks, which makes more sense anyway.